chore: App store dead code cleanup and restructuring  (#4497)
* moved chart-group in seperate code

* removed unused dependency

* removed dead code

* extracted resource tree

* moved notes

* resource movement

* removed unused code

* removed unused dependency

* commit methods

* extracted status update

* chore: clean up unused dead code

* updated: EA mode docker file

* updated: migration number

---------
