+++ pkg/appStore/chartGroup/repository/ChartGroupEntry.go
